More work on HTML Entity
Now both cases of HTML entity (name and numeric) are immplemented in Haskell.
Some work still needs to be done: moving remainder of code to Haskell, doing
TODOs in Haskell code. Additionally tests must be written.

The work listed above does not include updating the list of defined entity
names to HTML5. This will be done sometime in the far future.