Adding and using FFI code for HtmlTags module
C code for getting attribute value is now replaced with call to Haskell
function. Test code in C stays for now (but not for long).

FFI code is added to ffi sub-directory. This will allow me to better manage
"pure" Haskell code and separate it from FFI code. In particular the .hsc
files will be placed in the ffi sub-directory.