CssLength is now a struct (with int field), not a plain int
This change should help me in the future with reimplementing the CSS length
in a cleaner way: without encoding float and integer values + types through
bit operations in a single integer.

There is also a Length type in dw/ dir that is very similar to CssLength.
Perhaps one day they will be the same type.