Better implementation of parsing of CSS selectors
It looks better, and it feels better. And it produces better results (more
consistent with results of original C++ code).

Adding "whitespace" type of token because a whitespace character may be a
combinator. I had to add some duplicate functions in parser to generate and
handle the whitespace tokens.

The code still needs lots of cleanup.