Experimenting with alternative implementation of CssComplexSelector
The alternative complex selector type is a chain of interleaving Compound
Selectors and Combinators.

Parsing of compound selectors by the new code becomes more separated and
encapsulated, and should be easier to expand in the future to support other
simple selectors.

This is just an experiment. Maybe it will be successful, or maybe it will be
removed.