Css Matching code is now a pure Haskell code
After a lot of work in imperative/FFI code, I managed to rewrite matching
code for CSS Complex Selector in pure Haskell code. The code still resembles
the imperative form (one of functions is called Loop), but I plan to improve
the code.

Now that I have the code in Haskell, I will be able to refine the data type
used for Complex Selector and chose the optimal one. I hope to have type that
makes it impossible to represent invalid states, and at the same time a type
that makes writing recursive matching code easy.