Main loop of code parsing CSS has been moved to Haskell
A loop in cssparser.cc/parseCss() that was parsing all rulesets is now moved
to Haskell. This means that peeking of css context (with stylesheet) and
poking it is in FFI code is done only once per call to parseCss().

This greatly reduces execution time of the function.

Previously parsing of css for my test page (a local copy of soylentnews.org
page) took ~12.5 seconds. With this change the parsing takes only ~1.5
second. This means that the FFI operations (peek/poke) of css context were
very, very expensive.

Now I can re-implement parsing of CSS at-rules: @media and @import, but
@import may require adding some IO code.