Refactoring code related to doctree and doctree node
When the code will be completely migrated to Haskell, I won't be able to use
C pointers to build the doctree data structure. For now I'm replacing
'pointer as map key' with 'unique id as map key' to be able to build and use
the doctree data structure in Haskell.

Some work in this area is still needed.