Doctree push and pop copied to Haskell
Slowly starting to move doctree operations to Haskell. In this commit I'm
creating a global mutable structure that holds a doctree, and I'm adding push
and pop functions that modify the tree.

For now the push and pop functions exist in C++ and in Haskell, because there
are many places where the doctree is referenced in C++ code more or less
directly, but the goal is to replace all those references with Haskell
function calls.

In the end the doctree will exist entirely in Haskell. There will be no FFI
code for converting the doctree (or doctree nodes) between Haskell and C++.
Thanks to the global mutable structure the doctree will live in Haskell, and
only in Haskell.

Sooner or later the global structure will be replaced with something else,
but for now it should be kept. I don't want to write that much FFI code for
doctree data structures anymore, and it looks like I won't have to (just some
intermediaries for basic operations on tree).