Step closer to keeping CSS Contexts only in Haskell
The contexts can be now put into and updated in Haskell's global list. The
list is indexed by an integer/reference that can be used from C++ or Haskell
code.

In the near future the C++ code will only use the reference, it won't be
operating on context pointer. That way the css context data will exist only
in Haskell, and there won't be any need for FFI data conversions of CSS
Context data structure.

This work is not completed yet.

Similar work has been done recently for Doctree.