'user agent' style sheet is now entirely in Haskell
A css context with the style sheet is now a separate variable that is created
entirely in Haskell. A constructor function for new contexts can be made
simpler - it doesn't have to check a value of index (reference) of created
context: all contexts created with this ctor function are treated the same.