'string list' CSS value type is now really a list of strings
I need this change to avoid parsing a string of comma-separated tokens in CSS
style engine. This parsing has been already done in CSS parser.

This will break C++ a bit because some code may still expect to receive full
string of tokens, but sooner or later that code will also be rewritten in
Haskell.