Passing parent node's style attrs to styleEngineApplyStyleToGivenNode
The function will be able to use the attrs of parent node to process
inheritance of CSS values between parent/child nodes. I don't know when I
will implement such a feature, but it's good to be prepared for it in
advance.

The code already has been using parent node's font attrs, and now all of the
node's attrs can be used to process current node.