Fixing how declarations are compared when updating Declaration Set
After recent changes to constructors of CSS declarations the comparison of
the declarations needs to be fixed: I can't compare whole values (data
constructor + value passed to constructor), I need to compare only
constructors.

This diff is implementing this.

This problem has been found thanks to unit tests that I wrote for "add or
update" function.