BackgroundColor value constructor now takes a specific value
The value constructor no longer accepts a generic CssValue value, but a
"specialized" value of "specialized" type that can be used only in the
context of background color.

One of values of the new type is "inherit", which allows me to implement
inheritance of background color value.

I still don't plan to implement the two features (property-specific value
types and inheritance) for all css properties supported by hello, this is
just a demonstration that it's possible.