Some work around "color" and "background-color" properties
The "color" property is the second one that gets its own specialized type for
values. It is also the second property that supports "inherit" values.

The "inherit" value for "color" and "background-color" properties doesn't do
much yet because it seems that the two properties are inheritable by default,
so the "inherit" value doesn't bring anything new. To really demonstrate that
the value works for CSS properties, I will need to try implementing support
for "inherit" in properties that aren't inheritable by default.