Reworking how some border properties are parsed
The main reason behind this change was to avoid duplication of definition of
list of strings/enums. The list (e.g. "thin"/"medium"/"thick"/"inherit") was
present in two places: in Parser.hs, on top of the large lookup table, and in
Declaration.hs, in code translating from one of the strings to one of
declaration values. I wanted to get rid of the duplication of the list.

The change necessary to do that is not *that* big. If it was applied to
properties other than border-*, then the change could be ~25% of this diff.
But I did the change for the border-* properties, which meant that I also had
to rework handling of shortcut properties (technically I didn't have to, but
since I already started down this path...).

Adjusting other properties to use the same approach should be now simpler
since the basics are developed.

The main lookup table (from property name to constructor) will be simpler in
the future: the looked up value will contain only a constructor, without
additional list of enums or list of conversion functions. But this is still
ahead of us.