Removing tokensAsValueEnum funciton
The function is not used anymore, so I'm removing it.

Unit tests for the function are disabled for now, and will have to be
re-written for a function that replaces tokensAsValueEnum.

Refactoring few places in Declaration.hs that were parsing tokens as some
enum value for selected CSS property values: the definitions (key, value)
dictionaries (assoc lists) for some properties have been moved outside of
functions into top level. Perhaps some day the dictionaries will be
referenced in the big table in Parser.hs again.