Values of "content" CSS property now have dedicated Haskell type
This is the last property from the original code that had either enums or
parse functions defined in the list of properties. Right now the list of
properties contains only a string with name of the property and a pair of
constructors (old style constructor + new style constructor).

The enums where lists of strings that can be values of particular CSS
property.

Moving the enums (and parse functions) out of the table dictated by need to
have a single place where enums are defined. Otherwise I would have them
defined twice: once in the big list of properties, and again in mapping
between enum strings and Haskell values.