Starting to re-add support for "font" CSS shorthand property
I was going to add the support by using code used in parsing other shorthand
properties, but I decided to try a more comprehensive approach to parsing of
CSS values. I looked at the specification of allowed values of the property
and I started writing functions that call parsers in a way that allows
parsing values of different sub-properties occurring in different order,
zero, once or more time, etc.

So I ended up with some form of parser combinators :)

Two of them are unit tested. Unit tests for more combinators. Better
implementation of combinators is needed. But the direction has been set.