Improvements to Chain module
When fixing compilation warnings for the module (about some patterns not
being handled in function definitions), I noticed that the current definition
of the data type allows for tree-like layout of data in the Chain. This was
clearly not something that I wanted, so I have re-written the definition of
the type a bit.

I have also renamed few functions from the module.

And the compilation warnings have been fixed.

And the usage of the module's type and functions has been adjusted.