Global change of 'dillo' to 'hello'
Making a global change of project name.

In many places 'dillo' has been replaced with 'program', 'browser' or
'project' because 'hello' may not be a final name of the program.
'dillorc' is renamed to more generic 'browserrc'.

In other places, where a program/project name needs to be displayed, 'dillo'
is actually replaced by 'hello'.

'.dillo' dir (located in $HOME) is replaced by '.config/hello'.

A change of program version from 0.1.0 to 0.2.0 was made here too.