CSS: 'font' property is now in the main container of properties
The ctor for 'font' property is moved from container of shorthand properties
to container of all other properties.

The ctor is not updated to parse value of 'font' in new way. Currently the
value of CssPropertyFont is a list of non-shorthand font sub-properties
(non-shorthand properties) to which the shorthand expands. I will fix this later.

The 'font' property was a last shorthand property that was handled in
separate path. Now all CSS properties are looked up in the same map, and
there is a single place where property's constructor (parser) is applied.