CSS: code parsing a single declaration now returns Maybe
The code returned a list before because a single shorthand property could be
expanded to a list of non-shorthand properties. But after recent changes to
how shorthand properties are parsed (they are parsed into a record, not to a
list), I can now change list to Maybe in top-level functions parsing single
declaration.