CSS: new approach to parsing "font" shortcut property
I'm writing new implementation of my combinators and multipiers. The
combinators and multipliers now process the parsers of type
"Parser a (b -> b)".

The parser returns "b -> b" function that is a closure with value parsed by a
parser.

This may be an overkill, but this is the way that I want to investigate. It
allows me to use function composition in new context, and forces me to use
"flip" (which is an awesome function!).