CSS: deeper cleanup of cascading code that doesn't use match caching
Few elements that were needed for the match caching are now removed from
Cascade.hs.

This concludes the main part of simplifying of cascading. The cascading and
selector matching no longer use caching of results.

Still to do:
 - there are still some leftovers of caching throughout the code, they should
   be removed (caching in CSS context, cached complex selectors)
 - the code in selector matching needs to be reviewed: some control
   structures are no longer necessary.
 - check performance of cascading and selector matching. There may be some
   room for improvement in speed.

I acknowledge that there may be some necessity to speed up cascading or
selector matching, but at least now I can focus on learning how to do it in
"pure" Haskell code, and not in a poor rewrite of C code into Haskell code.
First get a good looking and correct code, then measure it, and then improve
its performance.