Updating code stats after removal of CSS match caching
Notice that this drop in count of Haskell lines is not due to me using smart
coding techniques. It is due to removal of code with which I don't want to
deal anymore.

I don't want to figure out how this particular piece of C++ code rewritten in
Haskell should be working or how to improve it - this is not what this
project is about. This project is about me learning Haskell.

Replacing a non-essential code (non-essential from point of view of web
browser) brings me closer to learning how to write a web browser in "pure"
Haskell.