Removing c_style_attrs_t
In the last tens of commits I have been removing fields of c_style_attrs_t
structure (and to some degree from StyleAttrs class). All the members of the
structure, i.e. the style attributes, are still present in Haskell code (in
StyleAttrs record), and they are available to C++ code through calls of FFI
functions.

There was now only one member of this struct left: an integer that serves as
a reference to Haskell object.

Since there was only one member, I removed the unused struct and used the
integer more directly.