Declaration sets are now kept in Haskell's Style Engine stack
Main (non-important), important and non-CSS declaration sets are kept only on
a stack of style nodes that are stored in Haskell code, in StyleEngine
record.

This means that code that processes and uses the declaration sets is entirely
in Haskell. C++ code accesses the declaration sets through integer reference,
and only in one place: after the declaration sets are merged. In that case
the integer reference is passed from one FFI function to another.

I noticed that some amount of non-trivial Haskell code is "hiding" in FFI
files. It will have to be moved out of FFI one day.