Removing doctree object from C++ code
The doctree now exists only in Haskell code. C++ code can access it only
through opaque integer reference.

The integer reference to the doctree existed for some time in code base in
parallel with the original solution using a C++ doctree object and a pointer
to the object, but now the pointer is no longer needed so it has been
removed.