CSS: clean up of import and media code leftovers
Removing some code (some of it already commented-out) related to import and
media rules.

After recent additions and changes in Haskell code for parsing of import and
media rules, I think that I have a good basis for continuation of handling of
import and media rules in Haskell. Therefore the old code from C++ can be
removed.